+++ Changelog.md
@@ -6,10 +6,12 @@ _yyyy.mm.dd_
 ### New features
 
 ### Changes
+* LearningAgent now receives the number of threads and registers from the LearningParameter.
 
 ### Bug fix
 * Fix non-determinism issue in ClassificationLearningAgent.
 * Fix intron detection when importing a dot file with TPGGraphDotImporter.
+
 
 ## Release version 0.2.0
 _2020.06.03_